requirements.txt
flask>=2.0.1
beautifulsoup4>=4.9.3
lxml>=4.6.3
requests>=2.26.0
mysql-connector-python>=8.0.26
python-dotenv>=0.19.0
